Delphes features adds

+++ part3/hbnb/app/api/v1/reviews.py
@@ -1,5 +1,7 @@
 from flask_restx import Namespace, Resource, fields, _http
 from app.services import facade
+from flask_jwt_extended import jwt_required, get_jwt_identity, get_jwt
+
 
 api = Namespace('reviews', description='Review operations')
 
@@ -8,7 +10,7 @@ review_model = api.model('Review', {
     'text': fields.String(required=True, description='Text of the review'),
     'rating': fields.Integer(required=True, description='Rating of the place'
                              '(1-5)'),
-    'user_id': fields.String(required=True, description='ID of the user'),
+    'user_id': fields.String(required=False, description='ID of the user'),
     'place_id': fields.String(required=True, description='ID of the place')
 })
 
@@ -29,6 +31,11 @@ review_response_model = api.model('ReviewResponse', {
                              description='Rating of the place (1-5)')
 })
 
+update_review_model = api.model('UpdateReview', {
+	'text': fields.String(required=False, description='New text of the review'),
+    'rating': fields.Integer(required=False, description='New rating of the place'
+                             '(1-5)')
+})
 
 @api.route('/')
 class ReviewList(Resource):
@@ -40,10 +47,46 @@ class ReviewList(Resource):
     @api.expect(review_model)
     @api.response(201, 'Review successfully created')
     @api.response(400, 'Invalid input data')
+    @api.response(404, 'Place not found')
+    @jwt_required()
     def post(self):
         """Register a new review"""
+        current_user_id = get_jwt_identity()
         # Automatically parses and validates request JSON
         review_data = api.payload
+        # Retrieve the place_id
+        place_id_to_review = review_data.get('place_id')
+        if not place_id_to_review:
+            api.abort(400, error='Missing place_id')
+            return {'error': 'Missing place_id'}, 400
+        # Retrieve the place to validate ownership
+        place_to_review = facade.get_place(place_id_to_review)
+        if not place_to_review:
+            api.abort(404, error=str(e))
+            return {'error': str(e)}, 404
+
+        # Check if the current user tries to review their own place
+        if place_to_review.owner_id == current_user_id:
+            api.abort(403, error='You cannot review your own place')
+            return {'error': 'You cannot review your own place'}, 403
+
+        # Check if the review already exists
+        existing_review = facade.get_review_by_place_and_user(
+            place_id_to_review, current_user_id)
+        if existing_review:
+            api.abort(403, error='You have already reviewed this place')
+            return {'error': 'You have already reviewed this place'}, 403
+
+        # Ensure user_id in payload matches authenticated user
+        if 'user_id' in review_data and \
+                review_data['user_id'] != current_user_id:
+            api.abort(403, error='You are not allowed to create a review for'
+                      'another user')
+            return {'error': 'You are not allowed to create a review for'
+                    'another user'}, 403
+
+        # Set user_id to the authenticated user's ID
+        review_data['user_id'] = current_user_id
         try:
             review = facade.create_review(review_data)
         except Exception as e:
@@ -88,13 +131,29 @@ class ReviewResource(Resource):
     @api.doc('Returns the updated review')
     @api.marshal_with(review_response_model, code=_http.HTTPStatus.OK,
                       description='Review updated successfully')
-    @api.expect(review_model)
+    @api.expect(update_review_model)
     @api.response(200, 'Review updated successfully')
     @api.response(400, 'Invalid input data')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def put(self, review_id):
         """Update a review's information"""
+        current_user_id = get_jwt_identity()
         review_data = api.payload
+        # Retrieve the review to validate ownership
+        review_to_update = facade.get_review(review_id)
+
+        # Check if the review exists
+        if not review_to_update:
+            api.abort(404, error=str(e))
+            return {'error': str(e)}, 404
+
+        # Check if the authenticated user is the creator of the review
+        if review_to_update.user_id != current_user_id:
+            api.abort(403, error=str(e))
+            return {'error': str(e)}, 403
+
         try:
             updated_review = facade.update_review(review_id,
                                                   review_data)
@@ -110,12 +169,28 @@ class ReviewResource(Resource):
     @api.doc('Deletes review')
     @api.response(200, 'Review deleted successfully')
     @api.response(404, 'Review not found')
+    @api.response(403, 'Unauthorized action')
+    @jwt_required()
     def delete(self, review_id):
         """Delete a review"""
+        current_user_id = get_jwt_identity()
+        # Retrieve the review to validate ownership
+        review_to_delete = facade.get_review(review_id)
+
+        # Check if the review already exists
+        if not review_to_delete:
+            api.abort(404, error=str(e))
+            return {'error': str(e)}, 404
+
+        # Check if the creator of the review is the current user
+        if review_to_delete.user_id != current_user_id:
+            api.abort(403, error=str(e))
+            return {'error': str(e)}, 403
+
         try:
             facade.delete_review(review_id)
         except Exception as e:
-            api.abort(404, str(e))
+            api.abort(404, error=str(e))
             return {'error': str(e)}, 404
         return f"Review {review_id} has been succesfully deleted", 200
